refactor map drawing logic into HealthRegionList component and add css styles for hovering over a region
