+++ client/src/components/HealthRegion.js
@@ -1,12 +1,10 @@
+import "./HealthRegion.css";
+
 /**
  *
  * @param {*} props => { pathData }
  * component takes in a SVG path data string and renders a health region
  */
 export default function HealthRegion(props) {
-  const style = {
-    fill: "blue",
-    stroke: "black",
-  };
-  return <path style={style} d={props.pathData} />;
+  return <path d={props.pathData} />;
 }
